agent-server/pyproject.toml
[tool.poetry]
name = "agent-server"
version = "0.1.0"
description = ""
authors = ["Lucheng_Hugh <hughqing@qq.com>"]
readme = "README.md"
packages = [
    { include = "agent_server", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
livekit-plugins-deepgram = ">=0.6.4"
livekit-plugins-openai = ">=0.8.0"
livekit-plugins-silero = ">=0.6.3"
python-dotenv = "^1.0.1"
langgraph = "^0.2.4"
langgraph-sdk = "^0.1.27"
matplotlib = "^3.9.2"
nest-asyncio = "^1.6.0"
debugpy = "^1.8.5"
livekit = ">=0.17.0"
pydantic = "^2.8.2"
livekit-agents = ">=0.10.2"
convex = "^0.6.0"
openapi-generator-cli = "^4.3.1"
convex_client = { path = "../convex-client", develop = true }
libs = { path = "../libs", develop = true }
agent-graph = { path = "../agent-graph", develop = true }
livekit-plugins-elevenlabs = ">=0.7.5"
elevenlabs = "^1.9.0"
python-debouncer = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
